# docs/referencia/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Applications Tracker API",
    "description": "\nAPI REST de **Applications Tracker**: seguimiento de solicitudes a puestos de trabajo.\n\n## Autenticación\n\nLa autenticación la gestiona SuperTokens (email y contraseña). Hay dos modos, y la\nAPI acepta los dos en todas las rutas protegidas:\n\n| Cliente | Modo | Cómo |\n|---|---|---|\n| Navegador (la SPA) | Cookies httpOnly | Login con la cabecera `st-auth-mode: cookie`. El navegador envía las cookies solo. |\n| Integraciones, scripts, Swagger | Bearer | Login con `st-auth-mode: header` (o sin esa cabecera). Se usa `Authorization: Bearer <st-access-token>`. |\n\n### Probar desde esta página\n\n1. Despliega **POST /auth/signin**, pulsa *Try it out* y ejecuta con un usuario existente.\n2. En *Response headers*, copia el valor de `st-access-token`.\n3. Pulsa **Authorize**, pega el token en *BearerAuth* y confirma.\n4. Las rutas con candado ya se pueden ejecutar. El token caduca a los **5 minutos**: renuévalo con `POST /auth/session/refresh` o repite el login.\n\n## Convenciones\n\n- **Errores de validación:** 422, con el detalle de cada campo en `detail`.\n- **Sin sesión:** 401 `{\"message\": \"unauthorised\"}`.\n- **Recursos de otro usuario:** 404, igual que si no existieran.\n- **Listados paginados:** `page` (desde 1) y `limit` (máximo 100) →\n  `{items, total, page, limit, pages}`.\n- **Fechas:** ISO 8601 en UTC.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/applications/{application_id}/interviews": {
      "get": {
        "tags": [
          "Interviews"
        ],
        "summary": "Listar las entrevistas de una solicitud",
        "description": "Entrevistas de la solicitud, ordenadas por fecha (RF-23).",
        "operationId": "list_interviews_api_v1_applications__application_id__interviews_get",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewRead"
                  },
                  "title": "Response List Interviews Api V1 Applications  Application Id  Interviews Get"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión, o el access token ha caducado o no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "No existe o pertenece a otro usuario (no se distingue).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Interviews"
        ],
        "summary": "Añadir una entrevista",
        "description": "Registra una entrevista de la solicitud (RF-40). No cambia el estado: si la\nsolicitud está en `applied` o `screening`, el frontend propone pasarla a\n`interviewing` usando `allowed_transitions` (RF-42), sin forzarlo.",
        "operationId": "create_interview_api_v1_applications__application_id__interviews_post",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewRead"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión, o el access token ha caducado o no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "No existe o pertenece a otro usuario (no se distingue).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Datos no válidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{application_id}/interviews/{interview_id}": {
      "patch": {
        "tags": [
          "Interviews"
        ],
        "summary": "Editar una entrevista",
        "description": "Actualización parcial (RF-40): solo cambian los campos enviados. Incluye\n`outcome`, que es como se registra el resultado (RF-41).",
        "operationId": "update_interview_api_v1_applications__application_id__interviews__interview_id__patch",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewRead"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión, o el access token ha caducado o no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "No existe o pertenece a otro usuario (no se distingue).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Datos no válidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Interviews"
        ],
        "summary": "Borrar una entrevista",
        "description": "Borra la entrevista (RF-40).",
        "operationId": "delete_interview_api_v1_applications__application_id__interviews__interview_id__delete",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Sin sesión, o el access token ha caducado o no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "No existe o pertenece a otro usuario (no se distingue).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Company not found"
            ]
          },
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "not_found"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "description": "Error de negocio. `code` es estable (contrato); `detail` es informativo."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterviewCreate": {
        "properties": {
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At",
            "description": "Fecha y hora de la entrevista."
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1440.0,
                "minimum": 1.0,
                "description": "Minutos."
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "interviewers": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Interviewers",
            "description": "Con quién.",
            "examples": [
              "Laura (RR. HH.)"
            ]
          },
          "interview_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewType"
              },
              {
                "type": "null"
              }
            ]
          },
          "format": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewFormat"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Máximo 5 000 caracteres."
          }
        },
        "type": "object",
        "required": [
          "scheduled_at"
        ],
        "title": "InterviewCreate"
      },
      "InterviewFormat": {
        "type": "string",
        "enum": [
          "online",
          "onsite",
          "phone"
        ],
        "title": "InterviewFormat"
      },
      "InterviewOutcome": {
        "type": "string",
        "enum": [
          "pending",
          "passed",
          "failed",
          "cancelled"
        ],
        "title": "InterviewOutcome",
        "description": "Resultado de la entrevista (RF-41). `PENDING` es el valor por defecto."
      },
      "InterviewRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "interviewers": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interviewers"
          },
          "interview_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewType"
              },
              {
                "type": "null"
              }
            ]
          },
          "format": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewFormat"
              },
              {
                "type": "null"
              }
            ]
          },
          "outcome": {
            "$ref": "#/components/schemas/InterviewOutcome"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "scheduled_at",
          "duration_minutes",
          "interviewers",
          "interview_type",
          "format",
          "outcome",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "InterviewRead"
      },
      "InterviewType": {
        "type": "string",
        "enum": [
          "screening",
          "technical",
          "hr",
          "cultural",
          "final",
          "other"
        ],
        "title": "InterviewType"
      },
      "InterviewUpdate": {
        "properties": {
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1440.0,
                "minimum": 1.0,
                "description": "Minutos."
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "interviewers": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Interviewers"
          },
          "interview_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewType"
              },
              {
                "type": "null"
              }
            ]
          },
          "format": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewFormat"
              },
              {
                "type": "null"
              }
            ]
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "InterviewUpdate",
        "description": "Actualización parcial: solo se modifican los campos enviados; `null` vacía un\ncampo opcional. Incluye `outcome`: es como se registra el resultado (RF-41)."
      },
...
      "UnauthorizedError": {
        "properties": {
          "message": {
            "type": "string",
            "const": "unauthorised",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "UnauthorizedError",
        "description": "Respuesta 401 de SuperTokens: no hay sesión, o el token ha caducado o es inválido."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "Access token obtenido en `POST /auth/signin` (cabecera de respuesta `st-access-token`). Caduca a los 5 minutos.",
        "scheme": "bearer"
      },
      "CookieAuth": {
        "type": "apiKey",
        "description": "Cookie httpOnly que usa el navegador tras un login con `st-auth-mode: cookie`. Swagger no puede fijarla a mano.",
        "in": "cookie",
        "name": "sAccessToken"
      }
    }
  },
...
}